Trekker ut felles kode for meta robots
